+++ src/vue-router/routes.js
@@ -2,6 +2,7 @@ export const vueRoutes = Object.freeze({
   app: { name: 'app' },
   auth: { name: 'auth' },
   login: { name: 'auth.login' },
+  profile: { name: 'app.profile' },
   willRequests: { name: 'app.will-requests' },
   willRequestsList: { name: 'app.will-requests.list' },
 })